faet/Updata Currency
## 재화변경
-재화가 int값을 넘어서 long으로 변경 및 오류수정

+++ Assets/01.Scripts/Data/CurrencyDate.cs
@@ -13,6 +13,6 @@ public class CurrencyInfo
     public CurrencyType type;
     public string currencyName;    // 재화 이름
     public Sprite icon;            // UI에 띄울 아이콘 이미지
-    public int initialAmount = 0;  //시작 시 기본 지급 재화량
+    public long initialAmount = 0;  //시작 시 기본 지급 재화량
 }
